chore: move test helper fns to own module
